service/requirements.txt
CherryPy==18.6.1
Flask==2.0.2
GitPython==3.1.26
Paste==3.5.0